Welcome Cons-Api

+++ v2/backend/pkg/commons/db/bigtable.go
@@ -18,6 +18,7 @@ import (
 	"github.com/gobitfly/beaconchain/pkg/commons/types"
 	"github.com/gobitfly/beaconchain/pkg/commons/utils"
 	itypes "github.com/gobitfly/eth-rewards/types"
+
 	"golang.org/x/sync/errgroup"
 	"google.golang.org/api/option"
 	"google.golang.org/protobuf/proto"
@@ -1918,7 +1919,7 @@ func (bigtable *Bigtable) GetValidatorEffectiveness(validators []uint64, epoch u
 	for validator, reading := range aggEffectiveness {
 		res = append(res, &types.ValidatorEffectiveness{
 			Validatorindex:        validator,
-			AttestationEfficiency: float64(reading.Sum) / float64(reading.Count) * 100,
+			AttestationEfficiency: reading.Sum / float64(reading.Count) * 100,
 		})
 	}
 
@@ -2652,7 +2653,7 @@ func GetCurrentDayClIncome(validator_indices []uint64) (map[uint64]int64, error)
 		return dayIncome, err
 	}
 
-	currentDay := uint64(lastDay + 1)
+	currentDay := lastDay + 1
 	startEpoch := currentDay * utils.EpochsPerDay()
 	endEpoch := startEpoch + utils.EpochsPerDay() - 1
 	income, err := BigtableClient.GetValidatorIncomeDetailsHistory(validator_indices, startEpoch, endEpoch)
